Refactor refreshes into RefreshService
